Make dedicated resolvers for project, owner, and study permissions
