+++ packages/server/schema.gql
@@ -28,24 +28,6 @@ A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date
 """
 scalar DateTime
 
-type UserModel {
-  id: ID!
-}
-
-type Permission {
-  user: UserModel!
-  role: Roles!
-  hasRole: Boolean!
-  editable: Boolean!
-}
-
-enum Roles {
-  OWNER
-  PROJECT_ADMIN
-  STUDY_ADMIN
-  CONTRIBUTOR
-}
-
 type TagSchema {
   dataSchema: JSON!
   uiSchema: JSON!
@@ -64,6 +46,16 @@ type Study {
   tagSchema: TagSchema!
   project: ID!
   tagsPerEntry: Float!
+}
+
+type UserModel {
+  id: ID!
+}
+
+type ProjectPermissionModel {
+  user: UserModel!
+  isProjectAdmin: Boolean!
+  editable: Boolean!
 }
 
 type Entry {
@@ -126,8 +118,7 @@ type Query {
   getOrganizations: [Organization!]!
   exists(name: String!): Boolean!
   getDatasets: [Dataset!]!
-  getProjectPermissions(project: ID!): [Permission!]!
-  getStudyPermissions(study: ID!): [Permission!]!
+  getProjectPermissions(project: ID!): [ProjectPermissionModel!]!
   projectExists(name: String!): Boolean!
   getProjects: [Project!]!
   studyExists(name: String!, project: ID!): Boolean!
@@ -145,8 +136,8 @@ type Mutation {
   createDataset(dataset: DatasetCreate!): Dataset!
   changeDatasetName(dataset: ID!, newName: String!): Boolean!
   changeDatasetDescription(dataset: ID!, newDescription: String!): Boolean!
-  grantOwner(targetUser: ID!): Boolean!
   grantProjectPermissions(project: ID!, user: ID!, isAdmin: Boolean!): Boolean!
+  grantOwner(targetUser: ID!): Boolean!
   signLabCreateProject(project: ProjectCreate!): Project!
   deleteProject(project: ID!): Boolean!
   createStudy(study: StudyCreate!): Study!
