+++ packages/server/src/auth/user.model.ts
@@ -1,0 +1,11 @@
+import { ObjectType, Field, Directive, ID } from '@nestjs/graphql';
+
+/** Definition for external user */
+@ObjectType()
+@Directive('@key(fields: "id")')
+@Directive('@extends')
+export class UserModel {
+  @Field(() => ID)
+  @Directive('@external')
+  id: string;
+}
