+++ packages/server/src/permission/models/project.model.ts
@@ -1,0 +1,15 @@
+import { ObjectType, Field } from '@nestjs/graphql';
+import { UserModel } from '../../auth/user.model';
+
+
+@ObjectType()
+export class ProjectPermissionModel {
+  @Field(() => UserModel)
+  user: string;
+
+  @Field()
+  isProjectAdmin: boolean;
+
+  @Field()
+  editable: boolean;
+}
