+++ packages/server/src/permission/permission.module.ts
@@ -1,14 +1,15 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { casbinProvider } from './casbin.provider';
-import { PermissionResolver } from './permission.resolver';
 import { PermissionService } from './permission.service';
 import { ProjectModule } from '../project/project.module';
 import { AuthModule } from '../auth/auth.module';
 import { StudyModule } from '../study/study.module';
+import { ProjectPermissionResolver } from './resolvers/project.resolver';
+import { OwnerPermissionResolver } from './resolvers/owner.resolver';
 
 @Module({
   imports: [forwardRef(() => ProjectModule), AuthModule, forwardRef(() => StudyModule)],
-  providers: [casbinProvider, PermissionResolver, PermissionService],
+  providers: [casbinProvider, PermissionService, ProjectPermissionResolver, OwnerPermissionResolver],
   exports: [casbinProvider]
 })
 export class PermissionModule {}
